Module aws_sdk_schemas::operation

Structs

CreateDiscoverer

Creates a discoverer.

CreateRegistry

Creates a registry.

CreateSchema

Creates a schema definition.

Inactive schemas will be deleted after two years.

DeleteDiscoverer

Deletes a discoverer.

DeleteRegistry

Deletes a Registry.

DeleteResourcePolicy

Delete the resource-based policy attached to the specified registry.

DeleteSchema

Delete a schema definition.

DeleteSchemaVersion

Delete the schema version definition

DescribeCodeBinding

Describe the code binding URI.

DescribeDiscoverer

Describes the discoverer.

DescribeRegistry

Describes the registry.

DescribeSchema

Retrieve the schema definition.

ExportSchema

GetCodeBindingSource

Get the code binding source URI.

GetDiscoveredSchema

Get the discovered schema that was generated based on sampled events.

GetResourcePolicy

Retrieves the resource-based policy attached to a given registry.

ListDiscoverers

List the discoverers.

ListRegistries

List the registries.

ListSchemaVersions

Provides a list of the schema versions and related information.

ListSchemas

List the schemas.

ListTagsForResource

Get tags for resource.

PutCodeBinding

Put code binding URI

PutResourcePolicy

The name of the policy.

SearchSchemas

Search the schemas

StartDiscoverer

Starts the discoverer

StopDiscoverer

Stops the discoverer

TagResource

Add tags to a resource.

UntagResource

Removes tags from a resource.

UpdateDiscoverer

Updates the discoverer

UpdateRegistry

Updates a registry.

UpdateSchema

Updates the schema definition

Inactive schemas will be deleted after two years.